requirements.txt
async-timeout==4.0.2
click==8.1.3
colorama==0.4.4
Deprecated==1.2.13
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
redis==4.3.2
six==1.16.0
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
